DOC: Avoid running examples twice

+++ doc/source/conf.py
@@ -74,35 +74,55 @@ def directory_size(directory_path):
 
 def remove_doctree(app, exception):
     """Remove the .doctree directory created during the documentation build."""
-    size = directory_size(app.doctreedir)
-    logger.info(f"Removing doctree {app.doctreedir} ({size} MB).")
-    shutil.rmtree(app.doctreedir, ignore_errors=True)
-    logger.info(f"Doctree removed.")
+    builder = app.builder.name
+
+    # Keep the examples already copied to avoid running them twice in CI
+    # when building html and pdf
+    if builder == "html" and os.getenv("ON_CI", False):
+        logger.info(f"Keeping directory {app.doctreedir}.")
+    else:
+        size = directory_size(app.doctreedir)
+        logger.info(f"Removing doctree {app.doctreedir} ({size} MB).")
+        shutil.rmtree(app.doctreedir, ignore_errors=True)
+        logger.info(f"Doctree removed.")
 
 
 def copy_examples(app):
     """Copy directory examples (root directory) files into the doc/source/examples directory."""
+    builder = app.builder.name
     destination_dir = pathlib.Path(app.srcdir, "examples").resolve()
-    logger.info(f"Copying examples from {EXAMPLES_DIRECTORY} to {destination_dir}.")
 
-    if os.path.exists(destination_dir):
-        size = directory_size(destination_dir)
-        logger.info(f"Directory {destination_dir} ({size} MB) already exist, removing it.")
-        shutil.rmtree(destination_dir)
-        logger.info(f"Directory removed.")
+    # Keep the examples already copied to avoid running them twice in CI
+    # when building html and pdf
+    if builder == "pdf" and os.getenv("ON_CI", False):
+        logger.info(f"Keeping directory {destination_dir}.")
+    else:
+        logger.info(f"Copying examples from {EXAMPLES_DIRECTORY} to {destination_dir}.")
 
-    shutil.copytree(EXAMPLES_DIRECTORY, destination_dir)
-    logger.info(f"Copy performed")
+        if os.path.exists(destination_dir):
+            size = directory_size(destination_dir)
+            logger.info(f"Directory {destination_dir} ({size} MB) already exist, removing it.")
+            shutil.rmtree(destination_dir)
+            logger.info(f"Directory removed.")
+
+        shutil.copytree(EXAMPLES_DIRECTORY, destination_dir)
+        logger.info(f"Copy performed")
 
 
 def remove_examples(app, exception):
     """Remove the doc/source/examples directory created during the documentation build."""
+    builder = app.builder.name
     destination_dir = pathlib.Path(app.srcdir) / "examples"
-    size = directory_size(destination_dir)
-    logger.info(f"Removing directory {destination_dir} ({size} MB).")
 
-    shutil.rmtree(destination_dir, ignore_errors=True)
-    logger.info(f"Directory removed.")
+    # Keep the examples result to avoid running them twice in CI
+    # when building html and pdf
+    if builder == "html" and os.getenv("ON_CI", False):
+        logger.info(f"Keeping directory {destination_dir}.")
+    else:
+        size = directory_size(destination_dir)
+        logger.info(f"Removing directory {destination_dir} ({size} MB).")
+        shutil.rmtree(destination_dir, ignore_errors=True)
+        logger.info(f"Directory removed.")
 
 
 # def add_ipython_time(app, docname, source):
@@ -283,7 +303,6 @@ extensions = [
     "sphinx.ext.autodoc",
     "sphinx.ext.todo",
     "sphinx.ext.autosummary",
-    "sphinx.ext.intersphinx",
     "sphinx.ext.coverage",
     "sphinx_copybutton",
     "sphinx_design",
@@ -292,7 +311,6 @@ extensions = [
     "recommonmark",
     "numpydoc",
     "ansys_sphinx_theme.extension.linkcode",
-    # "myst_parser"
 ]
 
 # MathJax config
